chore: separate eventParticipation
