+++ src/main/java/lems/cowshed/api/controller/event/EventSpecification.java
@@ -1,7 +1,6 @@
 package lems.cowshed.api.controller.event;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
 import lems.cowshed.api.controller.CommonResponse;
 import lems.cowshed.api.controller.ErrorCode;
 import lems.cowshed.api.controller.dto.event.response.*;
@@ -9,7 +8,7 @@ import lems.cowshed.api.controller.dto.event.request.EventSaveRequestDto;
 import lems.cowshed.api.controller.dto.event.request.EventUpdateRequestDto;
 import lems.cowshed.config.swagger.ApiErrorCodeExample;
 import lems.cowshed.config.swagger.ApiErrorCodeExamples;
-import lems.cowshed.service.CustomUserDetails;
+import lems.cowshed.domain.user.CustomUserDetails;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -38,21 +37,11 @@ public interface EventSpecification {
     CommonResponse<EventInfo> getEvent(@PathVariable("event-id") Long eventId,
                                        @AuthenticationPrincipal CustomUserDetails customUserDetails);
 
-    @Operation(summary = "모임 참여", description = "회원이 모임에 참여 합니다.")
-    @ApiErrorCodeExamples({ErrorCode.NOT_FOUND_ERROR})
-    CommonResponse<Void> saveEventParticipation(@Parameter(name="event-id", description = "모임 id", example = "1") @PathVariable Long eventId
-    , @AuthenticationPrincipal CustomUserDetails customUserDetails);
-
     @Operation(summary = "모임 수정", description = "모임의 세부 사항을 수정 합니다.")
     @ApiErrorCodeExamples({ErrorCode.NOT_FOUND_ERROR})
     CommonResponse<Void> editEvent(@PathVariable("event-id") Long eventId,
                                    @RequestBody @Validated EventUpdateRequestDto eventUpdateDto,
                                    @AuthenticationPrincipal CustomUserDetails customUserDetails);
-
-    @Operation(summary = "모임 참여 제거", description = "회원의 모임 참석을 제거 합니다.")
-    @ApiErrorCodeExamples({ErrorCode.NOT_FOUND_ERROR})
-    CommonResponse<Void> deleteEventParticipation(@Parameter(name="event-id", description = "모임 id", example = "1") @PathVariable Long eventId
-            , @AuthenticationPrincipal CustomUserDetails customUserDetails);
 
     @Operation(summary = "모임 삭제", description = "모임을 삭제 합니다.")
     @ApiErrorCodeExamples({ErrorCode.NOT_FOUND_ERROR})
